Refactor driver memory api.
