+++ src/OpenArk/OpenArk.vcxproj
@@ -253,6 +253,8 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\OpenArkDrv\arkdrv-api\api-memory\api-memory.cpp" />
+    <ClCompile Include="..\OpenArkDrv\arkdrv-api\api-storage\api-storage.cpp" />
     <ClCompile Include="..\OpenArkDrv\arkdrv-api\arkdrv-api.cpp" />
     <ClCompile Include="about\about.cpp" />
     <ClCompile Include="bundler\bundler.cpp" />
@@ -338,6 +340,8 @@
     </QtRcc>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\OpenArkDrv\arkdrv-api\api-memory\api-memory.h" />
+    <ClInclude Include="..\OpenArkDrv\arkdrv-api\api-storage\api-storage.h" />
     <ClInclude Include="..\OpenArkDrv\arkdrv-api\arkdrv-api.h" />
     <ClInclude Include="common\app\app.h" />
     <ClInclude Include="common\cache\cache.h" />
