+++ src/OpenArkDrv/arkdrv-api/api-storage/api-storage.h
@@ -1,0 +1,22 @@
+/****************************************************************************
+**
+** Copyright (C) 2019 BlackINT3
+** Contact: https://github.com/BlackINT3/OpenArk
+**
+** GNU Lesser General Public License Usage (LGPL)
+** Alternatively, this file may be used under the terms of the GNU Lesser
+** General Public License version 2.1 or version 3 as published by the Free
+** Software Foundation and appearing in the file LICENSE.LGPLv21 and
+** LICENSE.LGPLv3 included in the packaging of this file. Please review the
+** following information to ensure the GNU Lesser General Public License
+** requirements will be met: https://www.gnu.org/licenses/lgpl.html and
+** http://www.gnu.org/licenses/old-licenses/lgpl-2.1.html.
+**
+****************************************************************************/
+#pragma once
+#ifdef _ARKDRV_
+#include <ntifs.h>
+#include <windef.h>
+#else
+#include <Windows.h>
+#endif //_NTDDK_
